test(ui): align terminal continuation proof
